Fix issue 265
